virt-handler, filewatcher: move to separate top-level package
To be used in other places that extend beyond virt-handler.
