+++ pkg/filewatcher/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["filewatcher.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/virt-handler/filewatcher",
+    importpath = "kubevirt.io/kubevirt/pkg/filewatcher",
     visibility = ["//visibility:public"],
 )
