+++ pkg/virt-handler/multipath-monitor/BUILD.bazel
@@ -9,9 +9,9 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-handler/multipath-monitor",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/filewatcher:go_default_library",
         "//pkg/safepath:go_default_library",
         "//pkg/storage/reservation:go_default_library",
-        "//pkg/virt-handler/filewatcher:go_default_library",
         "//pkg/virt-handler/isolation:go_default_library",
         "//pkg/virt-handler/virt-chroot:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
@@ -28,8 +28,8 @@ go_test(
     race = "on",
     deps = [
         ":go_default_library",
+        "//pkg/filewatcher:go_default_library",
         "//pkg/safepath:go_default_library",
-        "//pkg/virt-handler/filewatcher:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
