+++ pkg/virt-handler/multipath-monitor/multipath_monitor_test.go
@@ -30,8 +30,8 @@ import (
 	. "github.com/onsi/gomega"
 	"go.uber.org/mock/gomock"
 
+	"kubevirt.io/kubevirt/pkg/filewatcher"
 	"kubevirt.io/kubevirt/pkg/safepath"
-	"kubevirt.io/kubevirt/pkg/virt-handler/filewatcher"
 	multipath_monitor "kubevirt.io/kubevirt/pkg/virt-handler/multipath-monitor"
 )
