+++ pkg/virt-handler/notify-server/BUILD.bazel
@@ -9,10 +9,10 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-handler/notify-server",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/filewatcher:go_default_library",
         "//pkg/handler-launcher-com/notify/info:go_default_library",
         "//pkg/handler-launcher-com/notify/v1:go_default_library",
         "//pkg/util/net/grpc:go_default_library",
-        "//pkg/virt-handler/filewatcher:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
